+++ modules/@apostrophecms/piece-type/index.js
@@ -128,7 +128,7 @@ module.exports = {
     return {
       add: {
         new: {
-          canEdit: true,
+          canCreate: true,
           relationship: true,
           label: {
             key: 'apostrophe:newDocType',
@@ -710,39 +710,6 @@ module.exports = {
       // True delete
       async delete(req, piece, options = {}) {
         return self.apos.doc.delete(req, piece, options);
-      },
-      composeFilters() {
-        self.filters = Object.keys(self.filters).map((key) => ({
-          name: key,
-          ...self.filters[key],
-          inputType: self.filters[key].inputType || 'select'
-        }));
-        // Add a null choice if not already added or set to `required`
-        self.filters.forEach((filter) => {
-          if (filter.choices) {
-            if (
-              !filter.required &&
-              filter.choices &&
-              !filter.choices.find((choice) => choice.value === null)
-            ) {
-              filter.def = null;
-              filter.choices.push({
-                value: null,
-                label: 'apostrophe:none'
-              });
-            }
-          } else {
-            // Dynamic choices from the REST API, but
-            // we need a label for "no opinion"
-            filter.nullLabel = 'Choose One';
-          }
-        });
-      },
-      composeColumns() {
-        self.columns = Object.keys(self.columns).map((key) => ({
-          name: key,
-          ...self.columns[key]
-        }));
       },
       // Enable inclusion of this type in sitewide search results
       searchDetermineTypes(types) {