+++ test/middleware-and-route-order.js
@@ -1,9 +1,9 @@
 const t = require('../test-lib/test.js');
 const assert = require('assert');
 
-let apos;
-
 describe('Middleware and Route Order', function() {
+
+  let apos;
 
   this.timeout(t.timeout);